Move DeprecatedCharacterSyntax to obsolescent
