+++ fortitude/resources/test/fixtures/obsolescent/OB062.f90
